Aligned native buffer wrapper
A utility buffer class to be used with JNA calls. Buffers are page
aligned (4k pages).

The wrapper mostly handles writes between ByteBuffers and
ByteBufs. It also provides a method for padding the buffer to the next
alignment, so writes can have an aligned size also (as required by
direct I/O). The padding is done with 0xF0, so that if it is read as
an integer, or long, the value will be negative (assuming the read is
a java read, and thus an signed int).