Direct I/O entry log writer implementation
The implementation uses JNA to do direct I/O to files via posix
syscalls. Fallocate is used if running on linux, otherwise this is
skipped (at the cost of more filesystem operates during writing).

There are two calls to write, writeAt and writeDelimited. I expect
writeAt to be used for the entrylog headers, which entries will go
through writeDelimited. In both cases, the calls may return before the
syscalls occur. #flush() needs to be called to ensure things are
actually written.

Direct I/O entry log reader implementation

Read from entrylog files using direct I/O posix calls.
Reads a block at a time, the size of the block determined by the size
of the buffers in the passed in buffer pool.

EntryLog ID generation for direct I/O

Generates unique entrylog ids. The implementation itself is trivial,
but it needs a decent amount of testing to ensure that it doesn't
stomp on preexisting entrylog id generation so that it is easy to move
between direct I/O and legacy.

Write path for direct entry logger

Writes entries to entry logs via DirectLogWriter, rolling the logs
once they reach a maximum size.

The entry log format isn't much changed from what is used by the
existing entrylogger. The biggest difference is the padding. Direct
I/O must write in aligned blocked. The size of the alignment varies by
machine configuration, but 4K is a safe bet on most. As it is unlikely
that entry data will land exactly on the alignment boundary, we need
to add padding to writes. The existing entry logger has been changed
to take this padding into account. When read as a signed int/long/byte
the padding will aways parse to a negative value, which distinguishes
it from valid entry data (the entry size will always be positive) and
also from preallocated space (which is always 0).

Another difference in the format is that the header is now 4K rather
than 1K. Again, this is to allow aligned rights. No changes are
necessary to allow the existing entry logger to deal with the header
change, as we create a dummy entry in the extra header space that the
existing entry logger already knows to ignore.

EntryLog index reading and log scanning

During garbage collection, the bookie needs to know how much of each
entry log is actually in use. Normally for this it will read the
index, but in the case that the index is absent, it scans the log.

This change implements both these for direct entrylogger.

Garbage Collection support for DirectIO entrylogger

GC support requires that the entrylogger provides a way to retrieve
all entrylogs which have been completely flushed to disk. Previously
this was done by returning the least unflushed log id. However, this
is problematic as it doesn't support the log ids wrapping around. It
also means that GC has to start checking for log id existence from
zero every time it boots.

This change replaces getLeastUnflushedLogId() with getFlushedLogIds(),
to give the entrylogger full control of which logs should be
considered for GC.

It also changes the CompactableLedgerStorage interface, removing
getEntryLogger() and adding injection of the entrylogger to the
GarbageCollectionThread. This makes testing easier.

Read direct I/O entrylog metadata in loop

Previously the metadata was not being read in a loop, which meant that
only the first batch of ledgers were being read in. This could have
caused data loss, as the metadata is used to decide if an entrylog
should be garbage collected, and if GC didn't see the full list of
ledgers in an entrylog, it could delete is prematurely.

Compaction support for direct I/O entry logger

Adds support for transactional compaction to the direct I/O entry
logger.
Adds an implementation of CompactionEntryLog. Most of the changes are
refactors to allow ComactionEntryLog to share functionallity with
DirectEntryLogger.

Wire up Direct I/O entry logger for db ledger storage

To enable, set dbStorage_directIOEntryLogger=true in the configuration.

Fix Direct IO entry log reader FD leak